Simplify by collapsing packages connection/ connectioncontext/ into networkservice/
Originally, connection and connectioncontext being separate from networkservice
were artifacts of the local/remote split in NSM.  This corrects that,
leaving us with two top level APIs:

1)  networkservice - Request/Close/MonitorConnection
2)  registry
