--- pkg/api/connection/gen.go
@@ -1,21 +1,0 @@
-// Copyright (c) 2020 Cisco Systems, Inc.
-//
-// SPDX-License-Identifier: Apache-2.0
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at:
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-// Package connection defines and provides helper methods for the Network Service Mesh API:
-//  Connection, Mechanisms, and MonitorConnection.
-package connection
-
-//go:generate bash -c "protoc -I . -I ../connectioncontext connection.proto --go_out=plugins=grpc,paths=source_relative:. --proto_path=$GOPATH/src/ --proto_path=$GOPATH/pkg/mod/  --proto_path=$( go list -f '{{ .Dir }}' -m github.com/golang/protobuf )"
