+++ pkg/api/networkservice/gen.go
@@ -17,4 +17,4 @@
 // Package networkservice provides the Network Service Mesh API: NetworkService{Server,Client}.{Request,Close}
 package networkservice
 
-//go:generate bash -c "protoc -I . -I ../connection -I ../connectioncontext networkservice.proto --go_out=plugins=grpc,paths=source_relative:. --proto_path=$GOPATH/src/ --proto_path=$GOPATH/pkg/mod/  --proto_path=$( go list -f '{{ .Dir }}' -m github.com/golang/protobuf )"
+//go:generate bash -c "protoc -I .  *.proto --go_out=plugins=grpc,paths=source_relative:. --proto_path=$GOPATH/src/ --proto_path=$GOPATH/pkg/mod/  --proto_path=$( go list -f '{{ .Dir }}' -m github.com/golang/protobuf )"
