+++ pkg/api/networkservice/mechanism_helpers.go
@@ -14,7 +14,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package connection
+package networkservice
 
 import (
 	"sync"
