+++ pkg/api/networkservice/mechanisms/srv6/constant.go
